+++ changelog.md
@@ -33,6 +33,8 @@
 - Add `--skip-proto` flag to `build`, `init` and `serve` commands to build the chain without building proto files
 - Add `node query tx` command to query a transaction in any chain.
 - Add `node query bank` command to query an account's bank balance in any chain.
+- Add `node tx bank send` command to send funds from one account to an other in any chain.
+- Add migration system for the config file to allow config versioning
 - Add `node tx bank send` command to send funds from one account to another in any chain.
 - Implement `network profile` command
 - Add `generate ts-client` command to generate a stand-alone modular TypeScript client.